Move StyleMixer files to scene directory
