+++ core/src/scene/styleMixer.cpp
